bcachefs: fix userspace build errors

+++ fs/bcachefs/btree_iter.c
@@ -2739,8 +2739,8 @@ void *__bch2_trans_kmalloc(struct btree_trans *trans, size_t size)
 {
 	struct bch_fs *c = trans->c;
 	unsigned new_top = trans->mem_top + size;
-	size_t old_bytes = trans->mem_bytes;
-	size_t new_bytes = roundup_pow_of_two(new_top);
+	unsigned old_bytes = trans->mem_bytes;
+	unsigned new_bytes = roundup_pow_of_two(new_top);
 	int ret;
 	void *new_mem;
 	void *p;
